requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-filter==2.4.0
idna==3.2
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.6
watchdog==2.1.3
websocket-client==1.2.1
